Make getDurationSeconds of long type

+++ src/main/java/io/iworkflow/core/command/TimerCommand.java
@@ -7,18 +7,18 @@ import java.time.Duration;
 @Value.Immutable
 public abstract class TimerCommand implements BaseCommand {
 
-    public abstract int getDurationSeconds();
+    public abstract long getDurationSeconds();
 
     public static TimerCommand createByDuration(String commandId, Duration duration) {
         return ImmutableTimerCommand.builder()
                 .commandId(commandId)
-                .durationSeconds((int) duration.getSeconds())
+                .durationSeconds(duration.getSeconds())
                 .build();
     }
 
     public static TimerCommand createByDuration(Duration duration) {
         return ImmutableTimerCommand.builder()
-                .durationSeconds((int) duration.getSeconds())
+                .durationSeconds(duration.getSeconds())
                 .build();
     }
 }
